+++ js/components/waqi.js
@@ -1,14 +1,14 @@
-/* global Dashticz */
+/* global Dashticz settings choose language */
 //# sourceURL=js/components/waqi.js
 (function (Dashticz) {
   "use strict";
   var DT_waqi = {
     name: 'waqi',
     canHandle: function (block) {
-      return block && false;
+      return block && block.type === 'waqi';
     },
     defaultCfg: function (block) {
-      var layout = choose(block.layout, 'large');
+      var layout = choose(block.layout, settings['waqi_layout'], 'large');
       var zoomfactors = {
         xsmall: [120, 1.2],
         small: [170, 0.9],
@@ -17,9 +17,11 @@
         xxl: [480, 1],
       }
       return {
+        icon: 'fas fa-wind',
+        title: 'Air Quality',
         width: 12,
         layout: layout, //xsmall, small, large, xlarge, xxl
-        city: 5771, //Amsterdam
+        city: choose(block.city, settings['waqi_city'], 5771), //Amsterdam
         refresh: 15*60,
         scaletofit: zoomfactors[layout][0] || 1,
         aspectratio: zoomfactors[layout][1] || 0.5,
@@ -33,15 +35,13 @@
         me.iframeid = me.mountPoint+'_iframe';
         html +=
           '<iframe style="border:0px' + height + ';" id="' + me.iframeid + '"></iframe>';
-    //      '<iframe is="x-frame-bypass"' + scrolling + ' style="border:0px' + height + ';"></iframe>';
         me.$mountPoint.find('.dt_state').html(html);
 
         var $iframe = me.$mountPoint.find('iframe');
         var $dtstate = me.$mountPoint.find('.dt_state');
         var width = parseInt(me.$mountPoint.find('div').innerWidth())  - (isXXL ? 20:0);
-//        var width = parseInt($dtstate.innerWidth()) - (isXXL ? 20:0);
         var scaling = me.block.scaletofit ? width/me.block.scaletofit : 1;
-        var iframeWidth = width/scaling; // - (isXXL?20:0);
+        var iframeWidth = width/scaling;
         var dtstatecss={};
         var iframecss={}
         var scalingStr = 'scale(' + scaling + ')';
@@ -56,40 +56,20 @@
         }
 
         if(me.block.layout==='xxl') {
-//          iframecss.width='100%';
-
             me.$mountPoint.find('.dt_block').addClass('xxl');
             $dtstate.css(dtstatecss);
-            $iframe.css(iframecss);    
+            $iframe.css(iframecss);
         }
         else {
-
           $dtstate.css(dtstatecss);
-          $iframe.css(iframecss);    
+          $iframe.css(iframecss);
         }
     },
-    /*
-    defaultContent: function (me) {
-        var scrolling =
-          me.block.scrollbars === false ||
-          navigator.userAgent.match(/(iPod|iPhone|iPad)/)
-            ? ' scrolling="no"'
-            : '';
-        var html = '';
-        var height = me.block.height ? ';height:' + me.block.height + 'px' : '';
-        me.iframeid = me.mountPoint+'_iframe';
-        html +=
-          '<iframe ' + scrolling + ' style="border:0px' + height + ';" id="' + me.iframeid + '"></iframe>';
-    //      '<iframe is="x-frame-bypass"' + scrolling + ' style="border:0px' + height + ';"></iframe>';
-    
-        return html;
-      },
-    */
     refresh: function(me) {
-        var iframe = '<script type="text/javascript" src="https://widgets.waqi.info/jswgt/?size=' + 
+        var iframe = '<script type="text/javascript" src="https://widgets.waqi.info/jswgt/?size=' +
         me.block.layout + '&city=@' + me.block.city +
         '"></script><noscript>' + language.misc.widget_not_visible + ' (<a href="https://aqicn.org/">' +
-        language.misc.more_info + '</a>)</noscript></iframe>';
+        language.misc.more_info + '</a>)</noscript>';
         var doc = document.getElementById(me.iframeid).contentWindow.document;
         doc.open();
         doc.write(iframe);